Organize the code structure in a more improved and professional way by adding folders and clearer division
