+++ app/infrastructure/db_connection.py
@@ -1,8 +1,6 @@
 import os
 import psycopg2 # type: ignore
-# from dotenv import load_dotenv
 
-# load_dotenv()
 DB_HOST = os.getenv("DB_HOST")
 DB_PORT = os.getenv("DB_PORT")
 DB_NAME = os.getenv("POSTGRES_DB")
