+++ app/routes/user_routes.py
@@ -1,6 +1,6 @@
 import http
 from flask import Blueprint, request
-from .user_service import UserService
+from services.user_service import UserService
 
 user_bp = Blueprint('user', __name__)
 user_service = UserService()
